[Page] update api.

+++ Page/Routes/api.php
@@ -11,10 +11,8 @@
 |
 */
 
-Route::group(['middleware' => ['auth:api']], function () {
-    Route::get('v1/page')->name('api.v1.page.index')->uses('Api\V1\PageController@index');
-    Route::post('v1/page')->middleware('permission:api.v1.page.*')->name('api.v1.page.index')->uses('Api\V1\PageController@store');
-    Route::get('v1/page/{page}')->name('api.v1.page.show')->uses('Api\V1\PageController@show');
-    Route::put('v1/page/{page}')->middleware('permission:api.v1.page.*')->name('api.v1.page.update')->uses('Api\V1\PageController@update');
-    Route::delete('v1/page/{page}')->middleware('permission:api.v1.page.*')->name('api.v1.page.destroy')->uses('Api\V1\PageController@destroy');
-});
+Route::get('v1/page')->name('api.v1.page.index')->uses('Api\V1\PageController@index');
+Route::post('v1/page')->middleware(['auth:api', 'permission:api.v1.page.*'])->name('api.v1.page.store')->uses('Api\V1\PageController@store');
+Route::get('v1/page/{page}')->name('api.v1.page.show')->uses('Api\V1\PageController@show');
+Route::put('v1/page/{page}')->middleware(['auth:api', 'permission:api.v1.page.*'])->name('api.v1.page.update')->uses('Api\V1\PageController@update');
+Route::delete('v1/page/{page}')->middleware(['auth:api', 'permission:api.v1.page.*'])->name('api.v1.page.destroy')->uses('Api\V1\PageController@destroy');
